fix some nit
